<area> without href should be treated as static text
https://www.w3.org/TR/html-aam-1.0/#html-element-role-mappings

The <area> tag should only be exposed as a Link with the Value Control
Pattern if it has an |href| attribute.

Adjusting the intrinsic role of AXImageMapLink to be Role::kStaticText
if it doesn't have a link.

Bug: 1040717
Change-Id: Icfa1c5e6d97f9bb8514df88b07b3768b36fc0d74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994525
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Adam Ettenberger <adettenb@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#730945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 008a534e0aea601aae92d5b52f809f51e6ff783f