v8bindings: NativeValueTraits<idl_interface> without code generation
Improves the implementation (template specialization) of
NativeValueTraits<_IDL_interface_> so that code generation
of template specialization of each concrete type won't be
necessary.

The binary size of libblink_*.so will be reduced as below:
  libblink_core.so     -261,768 bytes
  libblink_modules.so  -272,248 bytes
(component release build on GNU/Linux, x64)
Note that this is not the final binary size of the
executable to be shipped.  The effect on the shipped
binary should be much smaller.

Finally but not least importantly, this patch makes it
easier to migrate to the new bindings generator.

Bug: 839389
Change-Id: I61f517e539e06131aea8cf699ea5f2762275dee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1996105
Reviewed-by: Hitoshi Yoshida <peria@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fd83c8c1d8e0df23f13ebf67582765d9c42906e