Remove LayoutSearchField
It exists only to secure height for -webkit-search-cancel-button. The
code changes the intrinsic height of the INPUT box if the height of the
inner_editor box is shorter than the height of the
-webkit-search-cancel-button. However, it never happens because
-webkit-search-cancel-button's size alwyas matches to font-size, and we
ignore line-height property if the line-height is shoter than the
font-size.

So, we can remove LayoutSearchField, and this CL should have no behavior
changes.

Bug: 1040828
Change-Id: I7e35f662cffa757351044be0f3ddee8baaa27dc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1999886
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731168}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 962c3af952e67775a84c39240999e6f7b29d6d48