Reenable testing of allow-modals CSP directive.
https://crrev.com/217ff8855ced140af1f6f95f69991d98503858dc
removed the testing of the expected file because it was a passing
testharness.js file. However this was the test because the dialogs
get logged to output and compared in the -expected.txt file.

Resurrect the expected file and turn on logging of js dialogs for this
test case.

R=dcheng@chromium.org
BUG=1041588

Change-Id: I36ffabe05648a8b8c23c37b4e2e7387d256e1487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995474
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731337}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9a14518a8b8e5b064046a8d728318a5230a80f6