override new SkCanvas virtuals for transforms
- override new 4x4 virtual
- override scale/translate virtuals

In the new *brave* world, SkCanvas will call didTranslate() and didScale() in
response to the corresponding public calls. On the impl side, SkCanvas
no longer has a default impl for either, so they must be overridden.

Having to update multiple unittests might point to a slight problem with layering
and testing. It appears (to me) that most of the sites I had to update were
incidental, not core, to the test. The caller is baking in assumptions about how
Skia handles some high-level api calls (and what virtuals it may call). If we
could find a different way to still test the chrome/blink code in question, it
will be easier going forward.

bug: skia:9768

Change-Id: I8aa62feb354392105229ea45686b8208ba0e1450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998247
Commit-Queue: Mike Reed <reed@google.com>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731734}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32a1af6c27382ad01041ee19eb2d8f78db0f6dab