Time and datetime-local popup UI improvements
This CL updates the <input type="time"> and <input type="datetime-local"> popup UI's in three
related aspects per design feedback:

1) The in-page controls now update immediately as values are changed in the popup,
instead of waiting for the popup to be dismissed.  The page can watch for these 'live' changes
via the input event.
2) The UI is simplified by removal of the submit and cancel buttons.  The popups are now
dismissed only by clicking outside of them or with the Enter or Escape keys.
3) A 'double-escape' model is introduced in which if a user has opened the popup and changed the
value, hitting Escape once will reset the value back to what it was when the popup was opened.
Hitting Escape once more, when the popup value matches the value when opened, will dismiss the
popup.

The new live-update behavior required a change to how the platform fires change events for
temporal input popups.  Instead of firing it for every value setValue call--which would now
cause it to fire every time the user clicked a different value--it is now fired when the popup
is closed, via MultipleFieldsTemporalInputTypeView::DidEndChooser.  This is similar to how the
event is dispatched for the color input type popup.

Bug: 1001571
Change-Id: I464ba22768e4a722c1b6548a0128c849fbad6a39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988949
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#731979}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1cb29f13c9727fc3cb2e95a0bc42ce62ab1c1b0b