blink: Ensure correct shared image lifetime for static bitmap images.
AcceleratedStaticBitmapImages created by WebGL rely on creating a
GL texture backing for a shared image to ensure the shared image
outlives the StaticBitmapImage. This in turn assumes that even if the
client explicitly destroys a shared image mailbox via
SharedImageInterface, they can continue to use this mailbox as long as
an associated representation is alive. This pattern is not something we
want to support.

This pattern also means that when a mailbox is stored on
SharedImageInterface, we can not safely assume that the metadata
associated with it can be cached correctly.

This change fixes this by transitioning WebGL to use the same factory
method for creating StaticBitmapImages as used by canvas2d. This in
turn also does the following:

1) Ensures that the correct texture target is plumbed through with the
   StaticBitmapImage.
2) Eliminates the needs to plumb an image and a release callback through
   a bunch of code. Since the image now owns the release callback.
3) Cleans up code in MailboxTextureHolder used to keep the mailbox's
   texture backing alive.
4) TransferToImageBitmap currently works by marking the resource as
   lost, which means it can not be recycled. These redundant
   reallocations are avoided with this change.

R=kbr@chromium.org

Bug: 1036142
Change-Id: I8824aa8d03a48e35e39744cc30c9d0313dcbed08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1996124
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732100}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cc0504e52a9d7563b2ad51daec8faf87c719f53c