[video-raf] Add WebMediaPlayerMS::RequestAnimationFrame()
Callbacks passed to video.requestAnimationFrame are never called when
using webRTC. This is because webRTC uses WebMediaPlayerMS instead of
WebMediaPlayerImpl, and it does not override the RequestAnimationFrame()
method. This CL implements the method and fixes the issue.

Bug: 1035028, 1012063
Change-Id: I1792dfd6f6077e6e49c39f23e607e7868611293e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987069
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732265}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a6a928361e82aab0789d16205e0778cd32e1dd8