Use the portal-host-including parent for console logging in AncestorThrottle.
This is already used for determining whether ancestors are permitted.
It needs to be used also for figuring out which ancestor's console to
log in.

A web test that reaches this code using X-Frame-Options: DENY is included.
Without the change to content/, this test fails with the same stack trace
as in the linked bug.

Bug: 1038256
Change-Id: Iad3623d82e0b2192eab626a2184d09d9d3e7be62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002967
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 42f4df7a32818ccc22ce1199ab0a384388b0e923