[FlexNG] Miss cache if IntrinsicBlockSize was based on a fixed-block-size.
Previously within the cache logic if we had an indefinite initial
block-size we would compute a new final block-size based on the layout
results IntrinsicBlockSize().

For flex-boxes the IntrinsicBlockSize may depend on a previous
fixed-block-size (if present).

This resulted in an incorrect cache-hit.

A slightly more optimal version of this patch would be to set a bit on
this layout result named something like:
IntrinsicBlockSizeDependsOnInitialBlockSize().

This bit would be set if that layout had any child had its size based
on the initial-block-size. e.g. within the flex algorithm:

if (is_column_) {
  if (child.Style().ResolvedFlexGrow(Style()) > 0 ||
      child.Style().ResovledFlexShrink(Style()) > 0)
    container_builder_.SetIntrinsicBlockSizeDependsOnInitialBlockSize();
} else {
  if (DoesItemStretch(child))
    container_builder_.SetIntrinsicBlockSizeDependsOnInitialBlockSize();
}

... and within the cache logic:

if (layout_result.IntrinsicBlockSizeDependsOnInitialBlockSize())
  return NGLayoutCacheStatus::kNeedsLayout;

Bug: 845235
Change-Id: Iad4c5984ed894192a20d02303596d8fb9f71d036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002893
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732680}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4549a4ed1dcb81efb447115d586da0ae4de8c8f7