Don't set contents layer for <object> if it's using fallback content
Previously, when an <object> failed to load the remote frame content,
and used fallback content, we still got the cc::Layer of the remote
frame and set it as the contents layer of the GraphicsLayer of the
fallback content. When the fallback content had multiple GraphicsLayers
(e.g. when the object is inline containing block child), the
GraphicsLayers would reference the same cc layer, causing duplicated
cc layers in the layer list.

Now get cc layer of the remote frame only if the LayoutObject is
LayoutEmbeddedContent (i.e. not fallback).

Bug: 1033459
Change-Id: I95c1eb1e1a22f57e6ac12a2a10339230e3f4c90e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006970
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733009}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54969144db77b5a19a61547998d6488e81bc8d39