Guard TimingAllowPassed() usage on OutOfBlinkCors flag
Currently, the TimingAllowPassed() check is used regardless of whether
the OutOfBlinkCors flag is enabled or not. However, the check is only
computed when the flag is enabled, so in this CL we revert back to the
old behavior from before when the flag is disabled. See related change:
https://chromium-review.googlesource.com/c/chromium/src/+/1992590

Bug: 1003943, 1042580
Change-Id: Ifafbae93aaa58166fbb8007d30a3f18620029b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003754
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733029}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7a9012434c6bd66b8aca6755b2ac5cdaa14574e