[css-pseudo] Don't lay out ::marker if not originated by a list item
::marker pseudo-elements are originated by elements whose computed
'display' is 'list-item'. This usually means that they are list items,
that is, they are laid out as LayoutNGListItem (or LayoutListItem).

However, this is not always the case. E.g. a <fieldset> is laid out as
LayoutNGFieldset (or LayoutFieldset) even with 'display: list-item'.
This was the only problematic case in LayoutNG that I found in r731964,
so I just prevented <fieldset> from originating a ::marker. But this was
a fragile approach since there could be other cases. Actually I have
noticed other cases in legacy when working on implemententing markers
as pseudo-elements in legacy.

This patch adds a condition in PseudoElement::AttachLayoutTree to
prevent a ::marker from being laid out if it's originated by an element
which isn't laid out as a list item.

BUG=457718

Change-Id: I60fd774d6de79189b2871c98594d8bee397e4064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007457
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#733277}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73c5bc3f9edf4331e619574d4e12a029f186c80b