[LayoutNG] Remove IsIntermediateLayout flag.
Previously if we synthesized the min/max sizes by performing a layout
on the orthogonal subtree - we could end up in a situation where the
"DCHECK(!constraint_space.IsIntermediateLayout() || box_->NeedsLayout())"
was hit.

This was caused by a legacy layout fieldset needing a preferred widths
recalc.

This is an indirect fix for the below DCHECK bug. Removing the
IsIntermediateLayout flags means that we now perform a copy back when
we are in this (relatively rare) phase.

The IsIntermediateLayout was initially introduced as a performance
optimization - however it could also hurt performance if we perform
a correct layout of a subtree with a fixed size.
(we'd need to layout that sub-tree twice, getting exactly the same
layout result).

Bug: 1039241
Change-Id: I468a9c8e26b5e7855a72ddb2b1f2614538cedd21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007422
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733344}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4548e4da43ec9b5418574f8ae63db53c03ca4659