heap: ConcurrentMarking also handles write_barrier_worklist_
Restarting the markers only checked if the marking worklist is not empty
and ignored the write barrier worklist.
Also, moved the check so that it's physically next to the concurrent
marking method so hopefully if someone changes one in the future, they
will the next one right next ot it and change that too.

Bug: 986235
Change-Id: I8ce9dc66f0da4bfb1a4f0a37f56b3cdecdb0d0d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006351
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733590}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f742f8926bcbc5aa80b6bc4a6a75310672c96f9