heap: Bail out on concurrent marking for collections
HashTable (and types dependent on it), Vector and Deque are not yet
concurrent safe. If tracing on a concurrent thread, the collection
(and its trace method) is put in a separate marking worklist for
objects that are not safe to trace concurrently. This worklist is
processed by incremental marking on the mutator thread.

The bailout is explicitly called by classes that are not safe to trace
concurrently. At the start of a Trace method the bailout queries the
visitor whether it "supports" bailout. If it does, the visitor registers
the current object and Trace method. The mutator thread will later pick
up the object and recall the Trace method for it.

This entire mechanism is temporary so that we can enable concurrent
marking and get real world feedback without risk of data races. As
we make the collections concurrent safe, this mechanism will go away.

Binary size increase is also temporary. Binary size increase will
diminish as we move collections out of the bailout and enable
concurrent marking for them.

Bug: 986235
Change-Id: I713f293f79145dcfc51bd4d20b3e2e1a62dc415a
Binary-Size: Increase is temporary.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983221
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733626}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ce066a9af479aea7ef226cc0e247584aab874d6