Support non-zero paint offset for scroll bars
The previous code supposed that we created PaintOffsetTranslation for
all scrollable elements. This is not true for SVG foreign objects for
which we can't create PaintOffsetTranslation because the offset is in
SVG space which is affected by transform.

Bug: 1030109
Change-Id: I161821c50e091257f77b2f806ac5c00c903198da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012904
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733895}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3cd5609767a1a196e2fb288cfa289c2dbf4d16fd