Fix selected state when focused for ARIA gridcell and treeitem
Before this change, focused gridcell were perceived as selected by ATs.
This is because we assumed that a focused gridcell was also selected,
but that can't be the case.
Gridcells are mentioned here, but all ARIA grid roles and treeitem were
affected by this misinterpretation of the specs.

According to ARIA 1.0 and above, all ARIA grid cells, rows and columns
are selectable, but not necessarily selected. We can't assume a
relationship between the focused state and the selected state as they
are completely independent. We can only do so when the aria-selected
attribute is implied or required. To the best of my knowledge, it only
is implied and/or required for:
- ax::mojom::Role::kListBoxOption:
- ax::mojom::Role::kMenuListOption:
- ax::mojom::Role::kTab:

This CL introduces AXLayoutObject::SelectionShouldFollowFocus().
This method was implemented following the ARIA specs, so that only the
elements with an implied and/or required aria-selected attribute value
are also marked as selected when focused.

Bug: 1030378
Change-Id: I5875cf3f6198aeea73100806d365bb73f7a6c389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1996138
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#734001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4ccbed03152101811071c0ac04d73d75e2c86d09